crypto/requirements.txt
asn1crypto==1.5.1
bit==0.8.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coincurve==18.0.0
Flask==2.2.3
Flask-Admin==1.6.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==2.0.8
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
